% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy.R
\name{busy}
\alias{busy}
\title{Busy or not}
\usage{
busy(loader = "dots", position = "top right")
}
\arguments{
\item{loader}{character or list. If character given, then it selects from one of a few
available loaders (look Details for more). By using list you can create your custom loder.
List needs to have structure \code{loader <- list(css<-"CSS CODE", html <- div("CODE TO DISPLAY"))}.}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
shiny tag List with js script, style of spinner and display div
}
\description{
Displays a spinner indicating if there are any calculations running on the server side.
}
\details{
Currently available loaders: "dots", "spinner".

Solution inspired by: \url{https://colinfay.me/watch-r-shiny/}.
}
