% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display message on the top-right corner}
\usage{
display(message, position = "top right")
}
\arguments{
\item{message}{character with any message you want}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div which wraps your message to display it in the top-right corner of your shiny app.
}
\description{
Display message on the top-right corner
}
