% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{.key_condition}
\alias{.key_condition}
\title{Create a JS condition for a key}
\usage{
.key_condition(key, to_lower = TRUE)
}
\arguments{
\item{key}{key to convert into a condition}

\item{to_lower}{should the key be converted to lowercase. Default TRUE}
}
\value{
JS condition for checking shortcut
}
\description{
Create a JS condition for a key
}
\keyword{internal}
