% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy.R
\docType{data}
\name{.loader_dots}
\alias{.loader_dots}
\title{CSS/HTML code for a dots loader}
\format{An object of class \code{list} of length 2.}
\usage{
.loader_dots
}
\description{
CSS/HTML code for a dots loader
}
\keyword{datasets}
