% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy.R
\docType{data}
\name{.loader_spinner}
\alias{.loader_spinner}
\title{CSS/HTML code for a spinner loader}
\format{An object of class \code{list} of length 2.}
\usage{
.loader_spinner
}
\description{
CSS/HTML code for a spinner loader
}
\keyword{datasets}
