% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{.shortcut_condition}
\alias{.shortcut_condition}
\title{Create a JS condition for a shortcut}
\usage{
.shortcut_condition(shortcut)
}
\arguments{
\item{shortcut}{keys that should be checked. Shortcut can include special keys:
Ctrl, Alt, Shift. Keys should be separated with '+' sign. Default Ctrl+Shift+K}
}
\value{
JS condition for checking shortcut
}
\description{
Create a JS condition for a shortcut
}
\keyword{internal}
