% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{get_args}
\alias{get_args}
\title{Auxiliary function that returns list of arguments for parent calling function
See for details: https://stackoverflow.com/questions/17256834/getting-the-arguments-of-a-parent-function-in-r-with-names}
\usage{
get_args(fun = "infoPanel", depth = -9)
}
\arguments{
\item{fun}{string; for what function to look for}

\item{depth}{negative numeric; how layers above to check the arguments}
}
\value{
list of function arguments
}
\description{
Auxiliary function that returns list of arguments for parent calling function
See for details: https://stackoverflow.com/questions/17256834/getting-the-arguments-of-a-parent-function-in-r-with-names
}
