% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_info}
\alias{git_info}
\title{Git information}
\usage{
git_info(position = "top right")
}
\arguments{
\item{position}{character with position of the parameter. Default "top right".}
}
\description{
Displays git information from the repository of the current working space.
}
