% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{info_panel}
\alias{info_panel}
\title{Creates a panel for all info boxes so they do not overlap}
\usage{
info_panel(..., position = "top right")
}
\arguments{
\item{...}{calls with info elements}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div which wraps your all info boxes to display it in the position corner of your shiny app.
}
\description{
Creates a panel for all info boxes so they do not overlap
}
