% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{info_value}
\alias{info_value}
\title{UI output for info value}
\usage{
info_value(id, position = "top right")
}
\arguments{
\item{id}{output id to render}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div which wraps rendered value to display it in the {position} corner of your shiny app.
}
\description{
UI output for info value
}
