% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{infoPanel}
\alias{infoPanel}
\title{Creates a panel for all info boxes so they do not overlap}
\usage{
infoPanel(position = "top left", ...)
}
\arguments{
\item{position}{character with position of the parameter. Default "top right".}

\item{...}{calls with info elements}
}
\value{
div which wraps your all info boxes to display it in the {position} corner of your shiny app.
}
\description{
Creates a panel for all info boxes so they do not overlap
}
