% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_btn_server}
\alias{inspect_btn_server}
\title{Inspect button server}
\usage{
inspect_btn_server(input)
}
\arguments{
\item{input}{Shiny server input.}
}
\value{
observeEvent for Inspect button
}
\description{
Helps to easlily stop app at any moment for debugging or checking state.
}
