% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_btn_ui}
\alias{inspect_btn_ui}
\title{Inspect button UI}
\usage{
inspect_btn_ui(label = "Inspect", position = "top right")
}
\arguments{
\item{label}{label for debug button. Default "Inspect"}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div with "shinyinfo_inspect_btn".
}
\description{
Helps to easlily stop app at any moment for debugging or checking state.
}
