% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{link}
\alias{link}
\title{Link}
\usage{
link(label, url = "#", position = "top right")
}
\arguments{
\item{label}{character with the name of the link}

\item{url}{address of the link}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div with the link
}
\description{
Displays link to specified url
}
