% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{powered_by}
\alias{powered_by}
\title{Powered by}
\usage{
powered_by(
  company_name = NULL,
  link = "#",
  position = "top right",
  logo = NULL,
  logo_width = "120px",
  logo_height = "auto",
  inline = FALSE
)
}
\arguments{
\item{company_name}{character with the creator of the app}

\item{link}{link to the creator's website}

\item{position}{character with position of the parameter. Default "top right".}

\item{logo}{web link to a logo image or name of the image file in ./www}

\item{logo_width}{width of the logo in pixels}

\item{logo_height}{height of the logo in pixels}

\item{inline}{if TRUE, display text and logo on the same line}
}
\value{
div with "powered by".
}
\description{
Displays information about authors of the shiny app.
}
