% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{render_info_value}
\alias{render_info_value}
\title{Server render function for info value}
\usage{
render_info_value(expr, env = parent.frame(), quoted = FALSE,
  sep = " ")
}
\arguments{
\item{expr}{value to render}

\item{env}{The environment in which to evaluate expr. Default parent.frame()}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable. Default FALSE}

\item{sep}{A separator passed to cat to be appended after each element.}
}
\value{
Shiny render function to be save as an element of output.
}
\description{
Server render function for info value
}
\examples{
\dontrun{
# ui

info_value("value_to_display")

# server

test_reactive <- reactiveVal("some value")
output$value_to_display <- render_info_value(test_reactive())
}
}
