% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Display version of the app}
\usage{
version(ver = NULL, position = "top right")
}
\arguments{
\item{ver}{(default NULL) custom version number}

\item{position}{character with position of the parameter. Default "top right".}
}
\description{
Displays the version of the app by default using VERSION global variable.
}
\examples{
version() # if VERSION global variable exists
version("1.2.1") # with custom version number
}
