% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{infoPanel}
\alias{infoPanel}
\title{Wrapper for info_panel function}
\usage{
infoPanel(..., position = "top right")
}
\arguments{
\item{...}{calls with info elements}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div which wraps your all info boxes to display it in the position corner of your shiny app.
}
\description{
Wrapper for info_panel function
}
