% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{powered_by}
\alias{powered_by}
\title{Powered by}
\usage{
powered_by(company_name, link = "#", position = "top right")
}
\arguments{
\item{company_name}{character with the creator of the app}

\item{link}{link to the creator's website}

\item{position}{character with position of the parameter. Default "top right".}
}
\value{
div with "powered by".
}
\description{
Displays information about authors of the shiny app.
}
