% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{render_info_value}
\alias{render_info_value}
\title{Server render function for info value}
\usage{
output$test <- render_info_value("test value")
}
\arguments{
\item{expr}{value to render}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}

\item{sep}{A separator passed to cat to be appended after each element.}
}
\value{
Shiny render function to be save as an element of output.
}
\description{
Server render function for info value
}
